requirements.txt
aiofiles==23.2.1
anyio==4.1.0
argcomplete==3.2.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
boto3==1.34.2
botocore==1.34.2
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
dacite==1.7.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
fiftyone==0.23.1
fiftyone-brain==0.14.2
fiftyone-db==1.0
filelock==3.13.1
fonttools==4.46.0
fqdn==1.5.1
fsspec==2023.12.0
ftfy==6.1.3
future==0.18.3
glob2==0.7
graphql-core==3.2.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
humanize==4.9.0
Hypercorn==0.15.0
hyperframe==6.0.1
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.0
importlib-resources==6.1.1
inflate64==1.0.0
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonlines==4.0.0
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kaleido==0.2.1
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
mongoengine==0.24.2
motor==3.3.2
mpmath==1.3.0
multivolumefile==0.2.3
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
pprintpp==0.4.0
priority==2.0.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
py7zr==0.20.8
pybcj==1.0.2
pycocotools==2.0.7
pycparser==2.21
pycryptodomex==3.19.0
Pygments==2.17.2
pymongo==4.6.1
pyparsing==3.1.1
pyppmd==1.1.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
pyzstd==0.15.9
qtconsole==5.5.1
QtPy==2.4.1
rarfile==4.1
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
s3transfer==0.9.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.34.0
strawberry-graphql==0.138.1
sympy==1.12
tabulate==0.9.0
taskgroup==0.0.0a4
tenacity==8.2.3
terminado==0.18.0
texttable==1.7.0
threadpoolctl==3.2.0
tifffile==2023.12.9
tinycss2==1.2.1
tomli==2.0.1
torch==2.1.1
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
universal-analytics-python3==1.1.1
uri-template==1.3.0
urllib3==2.1.0
voxel51-eta==0.12.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.13.0
zipp==3.17.0
